refactor: rename routes for better understanding
