+++ app/api/cars/delete_duplicates/route.ts
